tock_registers::fields

Struct FieldValue

pub struct FieldValue<T: UIntLike, R: RegisterLongName> {
    pub value: T,
    /* private fields */
}

Values for the specific register fields.

For the FieldValue, the masks and values are shifted into their actual location in the register.

Fields
value: T
Implementations

impl<R: RegisterLongName> FieldValue<u8, R>

pub const fn new(mask: u8, shift: usize, value: u8) -> Self

impl<R: RegisterLongName> FieldValue<u16, R>

pub const fn new(mask: u16, shift: usize, value: u16) -> Self

impl<R: RegisterLongName> FieldValue<u32, R>

pub const fn new(mask: u32, shift: usize, value: u32) -> Self

impl<R: RegisterLongName> FieldValue<u64, R>

pub const fn new(mask: u64, shift: usize, value: u64) -> Self

impl<R: RegisterLongName> FieldValue<u128, R>

pub const fn new(mask: u128, shift: usize, value: u128) -> Self

impl<R: RegisterLongName> FieldValue<usize, R>

pub const fn new(mask: usize, shift: usize, value: usize) -> Self

impl<T: UIntLike, R: RegisterLongName> FieldValue<T, R>

pub fn none() -> Self

pub const fn mask(&self) -> T

Get the raw bitmask represented by this FieldValue.

pub fn read(&self, field: Field<T, R>) -> T

pub fn modify(self, val: T) -> T

Modify fields in a register value

pub fn matches_any(&self, val: T) -> bool

Check if any specified parts of a field match

pub fn matches_all(&self, val: T) -> bool

Check if all specified parts of a field match