tock_registers

Module registers

Implementation of included register types.

This module provides a standard set of register types, which can describe different access levels:

ReadWrite for registers which can be read and written to

ReadOnly for registers which can only be read

WriteOnly for registers which can only be written to

Aliased for registers which can be both read and written, but represent different registers depending on the operation

InMemoryRegister provide a register-type in RAM using volatile operations

These types can be disabled by removing the register_types crate feature (part of the default features). This is useful if this crate should be used only as an interface library, or if all unsafe code should be disabled.

Structs

Aliased

Read-only and write-only registers aliased to the same address.

InMemoryRegister

In memory volatile register.

ReadOnly

Read-only registers.

ReadWrite

Read/Write registers.

WriteOnly

Write-only registers.