tock_registers::registers

Struct Aliased

pub struct Aliased<T: UIntLike, R: RegisterLongName = (), W: RegisterLongName = ()> { /* private fields */ }

Read-only and write-only registers aliased to the same address.

Unlike the ReadWrite register, this represents a register which has different meanings based on if it is written or read. This might be found on a device where control and status registers are accessed via the same memory address via writes and reads, respectively.

This register implements Readable and Writeable, but in general does not implement ReadWriteable (only if the type parameters R and W are identical, in which case a ReadWrite register might be a better choice).