Trait tock_registers::interfaces::ReadWriteable

pub trait ReadWriteable {
    type T: UIntLike;
    type R: RegisterLongName;

    // Required method
    fn modify(&self, field: FieldValue<Self::T, Self::R>);
}

Readable and Writeable register, over the same RegisterLongName

Register which supports both reading and setting a value.

This trait does not have to be implemented manually! It is automatically implemented for every type that is both Readable and Writeable, as long as Readable::R == Writeable::R (i.e. not for Aliased registers).

Required Associated Types

type T: UIntLike

type R: RegisterLongName

Required Methods

fn modify(&self, field: FieldValue<Self::T, Self::R>)

Write the value of one or more fields, leaving the other fields unchanged