Struct tock_registers::LocalRegisterCopy

pub struct LocalRegisterCopy<T: UIntLike, R: RegisterLongName = ()> { /* private fields */ }

A read-write copy of register contents.

This behaves very similarly to a read-write register, but instead of doing a volatile read to MMIO to get the value for each function call, a copy of the register contents are stored locally in memory. This allows a peripheral to do a single read on a register, and then check which bits are set without having to do a full MMIO read each time. It also allows the value of the register to be “cached” in case the peripheral driver needs to clear the register in hardware yet still be able to check the bits. You can write to a local register, which will modify the stored value, but will not modify any hardware because it operates only on local copy.

This type does not implement the Readable and Writeable traits because it requires a mutable reference to modify the contained value. It still mirrors the interface which would be exposed by a type implementing Readable, Writeable and ReadWriteable.

Implementations

impl<T: UIntLike, R: RegisterLongName> LocalRegisterCopy<T, R>

pub const fn new(value: T) -> Self

pub fn get(&self) -> T

Get the raw register value

pub fn set(&mut self, value: T)

Set the raw register value

pub fn read(&self, field: Field<T, R>) -> T

Read the value of the given field

pub fn read_as_enum<E: TryFromValue<T, EnumType = E>>(
&self,
field: Field<T, R>
) -> Option<E>

Read value of the given field as an enum member

pub fn write(&mut self, field: FieldValue<T, R>)

Write the value of one or more fields, overwriting the other fields with zero

pub fn modify(&mut self, field: FieldValue<T, R>)

Write the value of one or more fields, leaving the other fields unchanged

pub fn is_set(&self, field: Field<T, R>) -> bool

Check if one or more bits in a field are set

pub fn matches_any(&self, field: FieldValue<T, R>) -> bool

Check if any specified parts of a field match

pub fn matches_all(&self, field: FieldValue<T, R>) -> bool

Check if all specified parts of a field match

pub fn bitand(&self, rhs: T) -> LocalRegisterCopy<T, R>

Do a bitwise AND operation of the stored value and the passed in value and return a new LocalRegisterCopy.